+++ modules/host/persist_compat_1.6.0_test.go
@@ -1,0 +1,1 @@
+package host
